+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Added `$billedCharacters` to the translate text response.
+
+
 ## [1.8.0] - 2024-06-24
 ### Added
 * Added document minification as a feature before document translation, to allow